pyproject.toml
[tool.poetry]
name = "nonebot-plugin-netrunner"
version = "0.0.1"
description = "a NoneBot plugin for Netrunner."
authors = ["Eric Li <eric03742@foxmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/eric03742/nonebot-plugin-netrunner"

[tool.poetry.dependencies]
python = "^3.9.0"
nonebot2 = "^2.3.3"
nonebot-adapter-onebot = "^2.4.5"
httpx = "^0.27.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
